requirements.txt
tensorflow
numpy
opencv-python-headless
Pillow
matplotlib
seaborn
scikit-learn
streamlit
